requirements.txt
PyYAML
uritemplate
Pygments
markdown
django
djangorestframework
django-filter
django-guardian
python-dotenv
requests